Add validateTextChannel validator
